backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0

scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.1
scipy==1.11.4

mysql-connector-python==8.1.0
SQLAlchemy==2.0.19
Flask-SQLAlchemy==3.0.5

python-dotenv==1.0.0
openpyxl==3.1.2
xlrd==2.0.1
tqdm==4.66.1

matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
dash==2.14.2
